Update postprocessor to recursive apply, and nodejs wrapper
Previously the sequence of postprocessors gets applied to only
the first level of transclusion, such that nested transclusions
get postprocessed as a single transcluded text. This makes post-
transclude tracking of source text impossible.

Here we apply postprocessing to all levels during transclusion,
also sending in the depth parameter (starting at 1), so source
information and global depth is traceable when desired.

Changes the nodejs library exported function to a wrapped fn
that converts the cljs object to #js