don't delete the ipc process handle to speed up repeat requests
the current ipc request function instantiates a new process handle for
I/O then immediately removes it with `delete-process` after it finishes.
This overhead substantially slows down repeat requests as shown in the
previous benchmark.

in this commit we save the process handle to a global and reuse or
recreate one on subsequent requests.

no additional process management system is introduced here, so currently
the handle must be manually destroyed when no longer needed.

on a fresh start of the socket server:

```sh
node xcl/build/node-content-server.js --bind socket
```

```emacs-lisp
(let ((ntimes 30)
      (time (current-time)))
  (dotimes
      (i ntimes)
    (plist-get
     (xcl-make-socket-jsonprc-request
      "echo"
      (list :protocol "TEST"
            :directive "foo"))
     :protocol))
  (insert (format "IPC: %.06f/s for %d requests"
                  (/ (float-time (time-since time))
                     ntimes)
                  ntimes)))
;; IPC: 0.001965/s for 30 requests
```